backend/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from fastapi.staticfiles import StaticFiles
from pydantic import BaseModel
from typing import Optional, Dict, Any
import uuid
import os
from dotenv import load_dotenv

from avatar_engine import AvatarEngine
from flow_controller import FlowController
from checkin_controller import CheckinController
from chatbot_integration import ChatbotIntegration

# Load environment variables
load_dotenv()

app = FastAPI(title="Vernacular Avatar Flight Booking", version="1.0.0")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class ChatRequest(BaseModel):
    query: str
    uid: Optional[str] = None
    language: str = "en"

class StepRequest(BaseModel):
    uid: str
    step: str
    user_input: Dict[str, Any]
    language: str = "en"

# Global instances
# Use ngrok URL if available, otherwise local IP
import os
ngrok_url = os.getenv('NGROK_URL', '')
if ngrok_url:
    avatar_engine = AvatarEngine(base_ip=ngrok_url)
    print(f"📹 Using ngrok for videos: {ngrok_url}")
else:
    import socket
    def get_local_ip():
        try:
            s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
            s.connect(("8.8.8.8", 80))
            ip = s.getsockname()[0]
            s.close()
            return ip
        except:
            return "localhost"
    server_ip = get_local_ip()
    avatar_engine = AvatarEngine(base_ip=server_ip)
sessions = {}

# Mount videos directory
video_path = os.path.join(os.path.dirname(__file__), "..", "videos")
if os.path.exists(video_path):
    app.mount("/videos", StaticFiles(directory=video_path), name="videos")

@app.post("/chat")
async def chat_endpoint(request: ChatRequest):
    """Main chat endpoint for multilingual chatbot"""
    try:
        uid = request.uid or str(uuid.uuid4())

        if uid not in sessions:
            sessions[uid] = {
                "flow_controller": FlowController(uid, request.language),
                "checkin_controller": CheckinController(uid, request.language),
                "chatbot": ChatbotIntegration(uid),
                "current_step": None,
                "language": request.language,
                "booking_data": {},
                "checkin_data": {},
                "in_avatar_flow": False,
                "in_checkin_flow": False
            }

        session = sessions[uid]
        session["language"] = request.language

        chatbot_response = await session["chatbot"].process_query(
            request.query,
            request.language,
            session["in_avatar_flow"]
        )

        if chatbot_response.get("trigger_avatar"):
            session["in_avatar_flow"] = True
            session["current_step"] = "welcome"
            return await session["flow_controller"].start_avatar_flow(
                request.query, 
                request.language
            )
        
        if chatbot_response.get("trigger_checkin"):
            session["in_checkin_flow"] = True
            session["current_step"] = "welcome_checkin"
            return await session["checkin_controller"].start_checkin_flow(
                request.query,
                request.language
            )

        return chatbot_response

    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/avatar-step")
async def process_avatar_step(request: StepRequest):
    """Process avatar booking step"""
    try:
        if request.uid not in sessions:
            raise HTTPException(status_code=404, detail="Session not found")

        session = sessions[request.uid]
        session["language"] = request.language
        session["booking_data"][request.step] = request.user_input

        response = await session["flow_controller"].process_step(
            request.step,
            request.user_input,
            request.language
        )

        session["current_step"] = response.get("next_step")
        return response

    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/checkin-step")
async def process_checkin_step(request: StepRequest):
    """Process avatar check-in step"""
    try:
        if request.uid not in sessions:
            raise HTTPException(status_code=404, detail="Session not found")

        session = sessions[request.uid]
        session["language"] = request.language
        session["checkin_data"][request.step] = request.user_input

        response = await session["checkin_controller"].process_step(
            request.step,
            request.user_input,
            request.language
        )

        session["current_step"] = response.get("next_step")
        return response

    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/avatar-video/{step}")
async def get_avatar_video(step: str, language: str = "en"):
    """Get avatar video URL for a specific step"""
    try:
        video_url = avatar_engine.get_video_url(step, language)
        return {
            "video_url": video_url,
            "step": step,
            "language": language
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/session/{uid}")
async def get_session_info(uid: str):
    """Get current session information"""
    if uid not in sessions:
        raise HTTPException(status_code=404, detail="Session not found")

    session = sessions[uid]
    return {
        "uid": uid,
        "current_step": session["current_step"],
        "in_avatar_flow": session["in_avatar_flow"],
        "booking_data": session["booking_data"],
        "language": session["language"]
    }

@app.delete("/session/{uid}")
async def clear_session(uid: str):
    """Clear session data"""
    if uid in sessions:
        del sessions[uid]
        return {"message": "Session cleared"}
    return {"message": "Session not found"}

@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return {"status": "healthy", "message": "Backend is running!"}

@app.post("/test-avatar")
async def test_avatar_trigger(request: ChatRequest):
    """Test endpoint to verify avatar trigger logic"""
    chatbot = ChatbotIntegration("test_uid")
    result = await chatbot.process_query(request.query, request.language)
    return {
        "query": request.query,
        "language": request.language,
        "result": result,
        "should_trigger": result.get("trigger_avatar", False)
    }

if __name__ == "__main__":
    import uvicorn
    print("🚀 Starting Vernacular Avatar Booking Backend...")
    if ngrok_url:
        print(f"📍 Server: {ngrok_url}")
    else:
        print(f"📍 Server: http://localhost:8000")
    print("📖 API docs: http://localhost:8000/docs")
    uvicorn.run("main:app", host="0.0.0.0", port=8000, reload=True)